un-memoize option-selector
